removing cast from board size to see if resolve compile issues on linux

+++ board.cpp
@@ -197,7 +197,9 @@ std::string Board::toString() {
   }
   return output;
 }
-int Board::getSize() { return (int)board.size(); }
+int Board::getSize() { 
+  
+  return board.size(); }
 
 Tile *Board::getTile(int row, int col) {
   Tile *tile = nullptr;
